rework: ObjectInfo::getAddress. (#61)
* add: Wrappers for H5Oget_info

There's four versions of the macro: one before the function was
versioned and then three versioned functions.

We wrap the version before versioning as h5o_get_info1, which makes it
easier to use specific versions of the function.

* add: ctor ObjecInfo(const Object&)

This allows us to move the `H5_info${v}_t` related logic into
a single class `ObjecInfo`.

* add: backwards compatible getAddress.

One of the differences between the various versions of H5O_info_t is
whether or not they have an address or a VOL token. By splitting
`getAddress` from ObjectInfo, we can continue to call H5Oget_info1 to
get the address.

* fix: ifdeffery for H5O_info_t raw_info.

The logic is that we always want `H5O_info1_t`. However, in very
old versions of HDF5 it was called `H5O_info_t` instead.

* remove: ObjectInfo::getAddress

The function was moved to Object::getAddress.